Merge revision 8124 from trunk: * Reopening the commit dialog after a failed commit did not always restore the state of all selected items.
Call commit on a directory which contains a modified file which is out of date. All other files in the directory should be unmodified, so that you do not need to uncheck any items. This will make TortoiseSVN do a recursive commit on the directory. After confirming the commit dialog you see that the file is out of date and the commit dialog re-opens. So far so good, but you will notice that the previously checked state of the file is not restored. The attached patch fixes this.


git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/branches/1.4.x@8125 154be670-7969-47af-9ab0-c6fb378d2295